requirements.txt
numpy
pygame
pillow
tqdm